% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{extract_aes_info}
\alias{extract_aes_info}
\title{Info about the aesthetic mappings of a plot.}
\usage{
extract_aes_info(p)
}
\arguments{
\item{p}{A ggplot object.}
}
\value{
A list with a \code{mapping} data frame and the \code{x} label.
}
\description{
Info about the aesthetic mappings of a plot.
}
\keyword{internal}
