% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{extract_geom_info}
\alias{extract_geom_info}
\title{Info about the geoms used by a plot.}
\usage{
extract_geom_info(p)
}
\arguments{
\item{p}{A ggplot object.}
}
\value{
A data frame with one row per layer.
}
\description{
Info about the geoms used by a plot.
}
\keyword{internal}
